Move to data_class.py file
